reorganized KKT linear system class for I+LowRank
